Upgrade: React->18, Next->13, and Chakra-UI->2
