+++ __tests__/components/home/portfolio.test.tsx
@@ -1,21 +1,8 @@
-import { render, screen } from "test-utils";
+import { render, screen, setDeviceWidth } from "test-utils";
 import userEvent from "@testing-library/user-event";
-import matchMediaPolyfill from "mq-polyfill";
 import { Portfolio } from "components/home";
 
 describe("<Portfolio />", () => {
-  beforeAll(() => {
-    matchMediaPolyfill(window);
-    window.resizeTo = function resizeTo(width, height) {
-      Object.assign(this, {
-        innerWidth: width,
-        innerHeight: height,
-        outerWidth: width,
-        outerHeight: height,
-      }).dispatchEvent(new this.Event("resize"));
-    };
-  });
-
   test("<Portfolio.Item /> renders a button and calls openFeature() on click", () => {
     const openFeature = jest.fn();
     const { container } = render(
@@ -77,7 +64,7 @@ describe("<Portfolio />", () => {
   test("<Portfolio.Button /> renders a button with text", () => {
     const { container } = render(<Portfolio.Button />);
 
-    expect(screen.getByRole(/button/i)).toBeInTheDocument();
+    expect(screen.getByRole("button")).toBeInTheDocument();
     expect(screen.getByText(/View/i)).toBeInTheDocument();
     expect(container.firstChild).toMatchSnapshot();
   });
@@ -85,7 +72,7 @@ describe("<Portfolio />", () => {
 
 describe("<Portfolio.Inner />", () => {
   test("<Portfolio.Inner /> renders with a border-bottom if viewport width < 768px and projectsCount is 1", () => {
-    window.resizeTo(767, 1000);
+    setDeviceWidth(767);
 
     const numberOfProjects = 1;
     const indexInArray = 0;
@@ -104,7 +91,7 @@ describe("<Portfolio.Inner />", () => {
   });
 
   test("<Portfolio.Inner /> renders without a border-bottom if viewport width < 768px, projectsCount is 2 and it is the last project", () => {
-    window.resizeTo(767, 1000);
+    setDeviceWidth(767);
 
     const numberOfProjects = 2;
     const indexInArray = 1;
@@ -123,7 +110,7 @@ describe("<Portfolio.Inner />", () => {
   });
 
   test("<Portfolio.Inner /> renders with a border-bottom if viewport width < 768px, projectsCount is 2 and it is not the last project", () => {
-    window.resizeTo(767, 1000);
+    setDeviceWidth(767);
 
     const numberOfProjects = 2;
     const indexInArray = 0;
@@ -142,7 +129,7 @@ describe("<Portfolio.Inner />", () => {
   });
 
   test("<Portfolio.Inner /> renders with a border-bottom if viewport width > 768px and projectsCount is < 3", () => {
-    window.resizeTo(768, 1000);
+    setDeviceWidth(768);
 
     const numberOfProjects = 2;
     const indexInArray = 0;
@@ -161,7 +148,7 @@ describe("<Portfolio.Inner />", () => {
   });
 
   test("<Portfolio.Inner /> renders without a border-bottom if viewport width > 768px, projectsCount is 3 and it is on the bottom row", () => {
-    window.resizeTo(768, 1000);
+    setDeviceWidth(768);
 
     const numberOfProjects = 3;
     const indexInArray = 2;
@@ -180,7 +167,7 @@ describe("<Portfolio.Inner />", () => {
   });
 
   test("<Portfolio.Inner /> renders with a border-bottom if viewport width > 768px, it is not in the first row, and it is the last one to be rendered before wrapping to a new row", () => {
-    window.resizeTo(768, 1000);
+    setDeviceWidth(768);
 
     const numberOfProjects = 5;
     const indexInArray = 3;
