+++ components/feature/index.ts
@@ -3,6 +3,5 @@ export * as Header from "./Header";
 export * as Links from "./Links";
 export * as Tech from "./Tech";
 export * as Description from "./Description";
-export { CarouselMobile } from "./CarouselMobile";
-export { CarouselDesktop } from "./CarouselDesktop";
+export * as Carousel from "./Carousel";
 export * as Descriptions from "./descriptions"
