+++ jest.setup.ts
@@ -1,1 +1,2 @@
 import "@testing-library/jest-dom";
+import "mock-match-media/jest-setup"
