+++ package.json
@@ -11,25 +11,26 @@
     "ci:test": "jest --ci"
   },
   "dependencies": {
-    "@chakra-ui/react": "^1.8.5",
-    "@emotion/react": "^11.1.5",
-    "@emotion/styled": "^11.3.0",
+    "@chakra-ui/react": "^2.8.1",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@vercel/analytics": "^1.0.2",
     "fp-ts": "^2.12.3",
-    "framer-motion": "^4.1.11",
-    "next": "^12.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.1",
+    "framer-motion": "^10.16.4",
+    "next": "^13.5.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "simplebar-react": "^2.3.6"
   },
   "devDependencies": {
-    "@next/eslint-plugin-next": "^12.3.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
+    "@next/eslint-plugin-next": "^13.5.4",
+    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^13.1.9",
+    "@types/jest": "^29.5.5",
     "@types/node": "^12.12.21",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
+    "@types/react": "^18.2.28",
+    "@types/react-dom": "^18.2.13",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.0",
     "babel-jest": "^27.0.6",
@@ -38,8 +39,9 @@
     "eslint-plugin-prettier": "^3.4.0",
     "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "jest": "^27.0.6",
-    "mq-polyfill": "^1.1.8",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "mock-match-media": "^0.4.3",
     "prettier": "^2.2.1",
     "typescript": "^4.5.5"
   },
