+++ utils/test-utils.tsx
@@ -1,7 +1,14 @@
 import { render, RenderOptions, RenderResult } from "@testing-library/react";
 import { ChakraProvider } from "@chakra-ui/react";
 import { ReactElement } from "react";
+import { setMedia } from "mock-match-media";
 import { ChildrenProps } from "types";
+
+export const setDeviceWidth = (width: number): void =>
+  setMedia({
+    width: `${width}px`,
+    type: "screen",
+  });
 
 const withProviders = ({ children }: ChildrenProps): ReactElement => (
   <ChakraProvider>{children}</ChakraProvider>
